Change shift count argument for << >> and >>> in Bitwise, etc to be a long (from int).
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@25916 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/src-java/x10/core/Byte.java
@@ -144,9 +144,9 @@ final public class Byte extends java.lang.Number implements StructI, java.lang.C
     public Byte $ampersand(java.lang.Object b, Type t) { return Byte.$box($value & ((Byte)b).$value); }
     public Byte $bar(java.lang.Object b, Type t) { return Byte.$box($value | ((Byte)b).$value); }
     public Byte $caret(java.lang.Object b, Type t) { return Byte.$box($value ^ ((Byte)b).$value); }
-    public Byte $left$G(int count) { return Byte.$box($value << count); }
-    public Byte $right$G(int count) { return Byte.$box($value >> count); }
-    public Byte $unsigned_right$G(int count) { return Byte.$box($value >>> count); }
+    public Byte $left$G(long count) { return Byte.$box($value << count); }
+    public Byte $right$G(long count) { return Byte.$box($value >> count); }
+    public Byte $unsigned_right$G(long count) { return Byte.$box($value >>> count); }
     // for X10PrettyPrinterVisitor.exposeSpecialDispatcherThroughSpecialInterface
     public byte $ampersand$B(java.lang.Object b, Type t) { return (byte) ($value & ((Byte)b).$value); }
     public byte $bar$B(java.lang.Object b, Type t) { return (byte) ($value | ((Byte)b).$value); }
